requirements.txt
absl-py==0.9.0
amqp==2.6.1
astor==0.8.1
billiard==3.6.3.0
boto3==1.14.53
botocore==1.17.53
cachetools==4.1.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
cmake==3.18.2.post1
cvlib==0.2.3
cycler==0.10.0
decorator==4.4.2
dlib==19.21.0
docutils==0.15.2
ffmpeg==1.4
Flask==1.1.1
Flask-API==2.0
fpdf==1.7.2
gast==0.2.2
gazpacho==0.9.4
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-efficientnets==0.1.7
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
kombu==4.6.11
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.3.1
moviepy==1.0.3
networkx==2.5
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.1
pandas==0.25.3
Pillow==7.2.0
proglog==0.1.9
progressbar==2.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
quote==1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tifffile==2020.8.25
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
